♻️ refactor: replace `VerifyUserInsigniaUseCase` with `GetUserUseCase` and throw `InsigniaNotIncludedError` if the user has the specified insignia
