+++ apps/server/src/rest/controllers/profile/users/VerifyUserInsigniaController.ts
@@ -2,7 +2,8 @@ import type { Controller, Http } from '@stardust/core/global/interfaces'
 import type { RestResponse } from '@stardust/core/global/responses'
 import type { InsigniaRole } from '@stardust/core/global/structures'
 import type { UsersRepository } from '@stardust/core/profile/interfaces'
-import { VerifyUserInsigniaUseCase } from '@stardust/core/profile/use-cases'
+import { GetUserUseCase } from '@stardust/core/profile/use-cases'
+import { InsigniaNotIncludedError } from '@stardust/core/profile/errors'
 
 export class VerifyUserInsigniaController implements Controller {
   constructor(
@@ -12,11 +13,13 @@ export class VerifyUserInsigniaController implements Controller {
 
   async handle(http: Http): Promise<RestResponse> {
     const account = await http.getAccount()
-    const useCase = new VerifyUserInsigniaUseCase(this.repository)
-    await useCase.execute({
-      userId: account.id as string,
-      insigniaRole: this.insigniaRole.value,
-    })
+    const useCase = new GetUserUseCase(this.repository)
+    const user = await useCase.execute({ userId: account.id })
+
+    if (user.insigniaRoles?.includes(this.insigniaRole.value)) {
+      throw new InsigniaNotIncludedError()
+    }
+
     return http.pass()
   }
 }
