+++ packages/core/src/profile/domain/errors/InsigniaAlreadyAcquiredError.ts
