+++ packages/core/src/profile/domain/errors/InsigniaNotIncludedError.ts
@@ -1,4 +1,4 @@
-import { NotAllowedError } from '../../global/domain/errors/NotAllowedError'
+import { NotAllowedError } from '../../../global/domain/errors/NotAllowedError'
 
 export class InsigniaNotIncludedError extends NotAllowedError {
   constructor() {
