+++ packages/core/src/profile/domain/errors/NotEnoughCoinsError.ts
