+++ packages/core/src/profile/domain/errors/ShopItemNotAcquiredError.ts
