+++ packages/core/src/profile/domain/errors/StarAlreadyUnlockedError.ts
