+++ packages/core/src/profile/domain/errors/UserEmailAlreadyInUseError.ts
