+++ packages/core/src/profile/domain/errors/UserNameAlreadyInUseError.ts
