+++ packages/core/src/profile/domain/errors/UserNotFoundError.ts
