+++ packages/core/src/profile/domain/errors/UserSocialAccountAlreadyInUseError.ts
