+++ packages/core/src/profile/domain/errors/index.ts
