+++ packages/core/src/profile/use-cases/CompleteChallengeUseCase.ts
@@ -1,7 +1,7 @@
 import { Id } from '#global/domain/structures/Id'
 import type { UseCase } from '#global/interfaces/UseCase'
 import type { UserDto } from '../domain/entities/dtos'
-import { UserNotFoundError } from '../errors'
+import { UserNotFoundError } from '../domain/errors'
 import type { UsersRepository } from '../interfaces'
 
 type Request = {
