+++ packages/core/src/profile/use-cases/tests/AcquireAvatarUseCase.test.ts
@@ -3,7 +3,7 @@ import { mock, type Mock } from 'ts-jest-mocker'
 import { Integer } from '#global/domain/structures/Integer'
 import { UsersFaker } from '#profile/domain/entities/fakers/UsersFaker'
 import { AvatarAggregatesFaker } from '#profile/domain/aggregates/fakers/AvatarAggregatesFaker'
-import { UserNotFoundError } from '#profile/errors/UserNotFoundError'
+import { UserNotFoundError } from '#profile/domain/errors/UserNotFoundError'
 import { AcquireAvatarUseCase } from '../AcquireAvatarUseCase'
 import type { UsersRepository } from '../../interfaces'
