+++ packages/core/src/profile/use-cases/tests/CalculateRewardForChallengeCompletionUseCase.test.ts
@@ -2,7 +2,7 @@ import { mock, type Mock } from 'ts-jest-mocker'
 
 import type { UsersRepository } from '#profile/interfaces/UsersRepository'
 import { Id } from '#global/domain/structures/Id'
-import { UserNotFoundError } from '#profile/errors/UserNotFoundError'
+import { UserNotFoundError } from '#profile/domain/errors/UserNotFoundError'
 import { UsersFaker } from '#profile/domain/entities/fakers/UsersFaker'
 
 import { CalculateRewardForChallengeCompletionUseCase } from '../CalculateRewardForChallengeCompletionUseCase'
