+++ packages/core/src/profile/use-cases/tests/CreateUserUseCase.test.ts
@@ -3,8 +3,8 @@ import { mock, type Mock } from 'ts-jest-mocker'
 import type { UsersRepository } from '#profile/interfaces/UsersRepository'
 import { UsersFaker } from '#profile/domain/entities/fakers/UsersFaker'
 import { Logical } from '#global/domain/structures/Logical'
-import { UserEmailAlreadyInUseError } from '#profile/errors/UserEmailAlreadyInUseError'
-import { UserNameAlreadyInUseError } from '#profile/errors/UserNameAlreadyInUseError'
+import { UserEmailAlreadyInUseError } from '#profile/domain/errors/UserEmailAlreadyInUseError'
+import { UserNameAlreadyInUseError } from '#profile/domain/errors/UserNameAlreadyInUseError'
 import { User } from '#profile/domain/entities/index'
 import { CreateUserUseCase } from '../CreateUserUseCase'
