+++ packages/core/src/profile/use-cases/tests/ObserveNewUnlockedAchievementsUseCase.test.ts
@@ -3,7 +3,7 @@ import { mock, type Mock } from 'ts-jest-mocker'
 import { AchievementsFaker } from '#profile/domain/entities/fakers/AchievementsFaker'
 import { UsersFaker } from '#profile/domain/entities/fakers/index'
 import type { UsersRepository, AchievementsRepository } from '#profile/interfaces/index'
-import { UserNotFoundError } from '#profile/errors/UserNotFoundError'
+import { UserNotFoundError } from '#profile/domain/errors/UserNotFoundError'
 import { IdFaker } from '#global/domain/structures/fakers/IdFaker'
 import { Achievement } from '#profile/domain/entities/Achievement'
 import { ObserveNewUnlockedAchievementsUseCase } from '../ObserveNewUnlockedAchievementsUseCase'
