+++ packages/core/src/profile/use-cases/tests/UnlockStarUse.test.ts
@@ -2,7 +2,7 @@ import { mock, type Mock } from 'ts-jest-mocker'
 
 import { Id } from '#global/domain/structures/Id'
 import { UsersFaker } from '#profile/domain/entities/fakers/UsersFaker'
-import { UserNotFoundError } from '#profile/errors/UserNotFoundError'
+import { UserNotFoundError } from '#profile/domain/errors/UserNotFoundError'
 import type { UsersRepository } from '#profile/interfaces/UsersRepository'
 import { UnlockStarUseCase } from '../UnlockStarUseCase'
