+++ packages/core/src/profile/use-cases/tests/VerifyUserInsigniaUseCase.test.ts
@@ -5,8 +5,8 @@ import type { UsersRepository } from '#profile/interfaces/UsersRepository'
 import { UsersFaker } from '#profile/domain/entities/fakers/UsersFaker'
 import { InsigniaRole } from '#global/domain/structures/InsigniaRole'
 import { Id } from '#global/domain/structures/Id'
-import { UserNotFoundError } from '#profile/errors/UserNotFoundError'
-import { InsigniaNotIncludedError } from '#profile/errors/index'
+import { UserNotFoundError } from '#profile/domain/errors/UserNotFoundError'
+import { InsigniaNotIncludedError } from '#profile/domain/errors/index'
 
 describe('Verify User Insignia Use Case', () => {
   let repository: Mock<UsersRepository>
@@ -28,7 +28,6 @@ describe('Verify User Insignia Use Case', () => {
         insigniaRole: InsigniaRole.createAsEngineer().value,
       }),
     ).rejects.toThrow(UserNotFoundError)
-  
   })
 
   it('should throw an error if the user does not have the insignia', async () => {
@@ -49,7 +48,7 @@ describe('Verify User Insignia Use Case', () => {
       insigniaRoles: [insignia.value],
     })
     repository.findById.mockResolvedValue(user)
-    
+
     expect(
       useCase.execute({
         userId: user.id.value,
